Fix terminal scrolling and terminalNotify
The original statement seems to be a bit too precise regarding the height. Any tiny lag or one pixel movement leads to the console not being scrolled to the bottom anymore. Same applies for manually scrolling down which does not hide the `terminalNotify` properly. 

A bit larger "buffer" fixes this.